+++ HISTORY.md
@@ -2,4 +2,5 @@
 
 ## 0.1.0 (???)
 
+- Add in basic CLI with search & apply commands. (See [#2](https://github.com/zillow/battenberg-gitlab/pull/2))
 - Add in initial search and apply functionality. (See [#1](https://github.com/zillow/battenberg-gitlab/pull/1))